requirements.txt
playwright==1.43.0
pandas==1.5.3
numpy==1.23.5
python-dotenv==1.0.1
